requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.2.2
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.3
Werkzeug==2.2.2
zipp==3.11.0
Flask-PyMongo
flask_restful
Flask-Mail
PyJWT~=1.4.2
